requirements.txt
selenium
python-dotenv
aiohttp
web3
